Adder User's Score to Leaderboard + Rank

+++ app/src/main/java/com/example/nightdash/retrofitcoil/scoreDataClass.kt
@@ -2,5 +2,5 @@ package com.example.nightdash.retrofitcoil
 
 data class scoreDataClass(
     val name: String,
-    val score: Int
+    val score: Float
 )
